Clean up nav / put in own file
